requirements.txt
python-telegram-bot==20.6
yt-dlp==2023.10.13
aiohttp==3.8.6
python-dotenv==1.0.0